# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Castle Grooves Dashboard API
  version: 1.0.0
  description: Same-origin API for the Castle Grooves React dashboard.
servers:
  - url: /
paths:
...
  /auth/discord/callback:
    get:
      operationId: finishDiscordLogin
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses: { '302': { description: Session created and redirected to dashboard }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' } }
  /auth/logout:
    post:
      operationId: logout
      responses: { '204': { description: Session cleared }, '403': { $ref: '#/components/responses/Error' } }
  /api/v1/me:
    get:
      operationId: getMe
      responses: { '200': { description: Session user, content: { application/json: { schema: { type: object, required: [user, contractVersion], properties: { user: { $ref: '#/components/schemas/SessionUser' }, contractVersion: { const: 1.0.0 } } } } } }, '401': { $ref: '#/components/responses/Error' } }
  /api/v1/state:
    get:
      operationId: getState
      responses: { '200': { $ref: '#/components/responses/State' }, '401': { $ref: '#/components/responses/Error' } }
  /api/v1/voice-channels:
    get:
      operationId: listVoiceChannels
      description: Voice channels in the configured guild where the bot can connect and speak.
      responses: { '200': { description: Available voice channels, content: { application/json: { schema: { type: object, required: [channels], properties: { channels: { type: array, items: { $ref: '#/components/schemas/VoiceChannel' } } } } } } }, '401': { $ref: '#/components/responses/Error' } }
  /api/v1/player/actions:
    post:
      operationId: playerAction
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PlayerAction' } } } }
      responses: { '200': { $ref: '#/components/responses/State' }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /api/v1/search:
    get:
      operationId: search
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 1, maxLength: 200 }, example: daft punk harder better }
        - { name: source, in: query, schema: { enum: [ytsearch, ytmsearch, scsearch] } }
      responses: { '200': { description: Search results, content: { application/json: { schema: { type: object, required: [results], properties: { results: { type: array, items: { $ref: '#/components/schemas/SearchTrack' } } } } } } }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' } }
  /api/v1/queue/items:
    post:
      operationId: addQueueItem
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [query], properties: { query: { type: string, minLength: 1, maxLength: 500 }, voiceChannelId: { $ref: '#/components/schemas/SelectedVoiceChannelId' } } }, example: { query: "https://www.youtube.com/watch?v=dQw4w9WgXcQ" } } } }
      responses: { '201': { $ref: '#/components/responses/State' }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /api/v1/queue/enqueues:
    post:
      operationId: enqueueHistoryPlays
      description: Admit exact history recordings asynchronously. The optional voice channel selection is part of the idempotent intent.
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [operationId, issuedAt, instanceId, expectedQueueId, placement, items], properties: { operationId: { type: string, format: uuid }, issuedAt: { type: string, format: date-time }, instanceId: { type: string, format: uuid }, expectedQueueId: { type: [string, 'null'], format: uuid }, placement: { enum: [append, next] }, afterQueueItemId: { type: [string, 'null'], format: uuid }, expectedQueueRevision: { type: integer, minimum: 0 }, voiceChannelId: { $ref: '#/components/schemas/SelectedVoiceChannelId' }, items: { type: array, minItems: 1, maxItems: 50, items: { type: object, required: [clientItemId, playId], properties: { clientItemId: { type: string, maxLength: 64 }, playId: { type: string } } } } } } } } }
      responses: { '200': { description: Existing operation receipt }, '202': { description: Resolving operation receipt }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /api/v1/queue/items/{queueItemId}:
    delete:
      operationId: removeQueueItem
      parameters: [ { name: queueItemId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '200': { $ref: '#/components/responses/State' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' } }
  /api/v1/history/queue:
    post:
      operationId: addHistoryQueueBatch
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [queries], properties: { queries: { type: array, minItems: 1, maxItems: 25, items: { type: string, minLength: 1, maxLength: 500 } }, voiceChannelId: { $ref: '#/components/schemas/SelectedVoiceChannelId' } } } } } }
      responses: { '201': { $ref: '#/components/responses/State' }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /api/v1/queue:
    patch:
      operationId: patchQueue
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/QueuePatch' } } } }
      responses: { '200': { $ref: '#/components/responses/State' }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' }, '403': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /api/v1/history:
    get:
      operationId: history
      parameters:
        - { name: range, in: query, schema: { enum: [24h, weekly, monthly, yearly], default: monthly } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses: { '200': { description: Listening history, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/HistoryItem' } } } } } } }, '401': { $ref: '#/components/responses/Error' }, '503': { $ref: '#/components/responses/Error' } }
  /api/v1/history/plays:
    get:
      operationId: historyPlays
      parameters:
        - { name: range, in: query, schema: { enum: [24h, weekly, monthly, yearly], default: monthly } }
        - { name: timezone, in: query, schema: { type: string, default: America/Toronto } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string, maxLength: 200 } }
        - { name: requesterId, in: query, schema: { type: string } }
        - { name: weekday, in: query, schema: { type: integer, minimum: 0, maximum: 6 } }
        - { name: hourFrom, in: query, schema: { type: integer, minimum: 0, maximum: 23 } }
        - { name: hourTo, in: query, schema: { type: integer, minimum: 0, maximum: 24 }, description: Exclusive local end hour; 24 means midnight. }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses: { '200': { description: Filtered recorded track starts in a bounded page, content: { application/json: { schema: { type: object, required: [timezone, range, items, nextCursor], properties: { timezone: { type: string }, range: { type: string }, nextCursor: { type: [string, 'null'] }, items: { type: array, items: { $ref: '#/components/schemas/HistoryPlay' } } } } } } }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' }, '503': { $ref: '#/components/responses/Error' } }
  /api/v1/history/insights:
    get:
      operationId: historyInsights
      parameters:
        - { name: range, in: query, schema: { enum: [24h, weekly, monthly, yearly], default: monthly } }
        - { name: timezone, in: query, schema: { type: string, default: America/Toronto } }
      responses: { '200': { description: "Range-wide history signals from up to the 1,000 most recent recorded starts", content: { application/json: { schema: { $ref: '#/components/schemas/HistoryInsights' } } } }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' }, '503': { $ref: '#/components/responses/Error' } }
  /api/v1/history/room-picks:
    get:
      operationId: historyRoomPicks
      parameters:
        - { name: range, in: query, schema: { enum: [24h, weekly, monthly, yearly], default: monthly } }
        - { name: timezone, in: query, schema: { type: string, default: America/Toronto } }
      responses: { '200': { description: "Up to three repeated, tailored song picks for every human currently in the active room", content: { application/json: { schema: { $ref: '#/components/schemas/HistoryRoomPicks' } } } }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' }, '503': { $ref: '#/components/responses/Error' } }
  /api/v1/history/plays/{playId}/context:
    get:
      operationId: historyPlayContext
      parameters:
        - { name: playId, in: path, required: true, schema: { type: string } }
        - { name: range, in: query, schema: { enum: [24h, weekly, monthly, yearly], default: monthly } }
        - { name: timezone, in: query, schema: { type: string, default: America/Toronto } }
        - { name: playedAt, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: radius, in: query, schema: { type: integer, minimum: 1, maximum: 5, default: 2 } }
      responses: { '200': { description: "Recorded starts around an anchor, in chronological order", content: { application/json: { schema: { $ref: '#/components/schemas/HistoryPlayContext' } } } }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '503': { $ref: '#/components/responses/Error' } }
components:
  responses:
    Error: { description: Stable error envelope, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' }, example: { error: { code: FORBIDDEN, message: DJ permission is required., requestId: 018fc4d4-9348-7a09-ae12-4c858fd8a2b9 } } } } }
    State: { description: Latest authoritative dashboard state, content: { application/json: { schema: { type: object, required: [state], properties: { state: { $ref: '#/components/schemas/DashboardState' } } } } } }
  schemas:
    VoiceChannel: { type: object, required: [id, name], properties: { id: { type: string }, name: { type: string } } }
...
    SessionUser: { type: object, required: [id, username, avatarUrl, role], properties: { id: { type: string }, username: { type: string }, avatarUrl: { type: [string, 'null'], format: uri }, role: { enum: [admin, dj, viewer] } } }
...
    PlayerAction:
      oneOf:
        - { type: object, required: [type], properties: { type: { const: pause } } }
        - { type: object, required: [type], properties: { type: { const: resume } } }
        - { type: object, required: [type], properties: { type: { const: skip } } }
        - { type: object, required: [type, disconnect], properties: { type: { const: stop }, disconnect: { const: true } } }
        - { type: object, required: [type, positionMs], properties: { type: { const: seek }, positionMs: { type: integer, minimum: 0 } } }
        - { type: object, required: [type, volume], properties: { type: { const: setVolume }, volume: { type: integer, minimum: 0, maximum: 200 } } }
        - { type: object, required: [type, repeatMode], properties: { type: { const: setRepeatMode }, repeatMode: { enum: [off, track, queue] } } }
...